style DelegateCardList, minimum prettier tabWidth 2 config and setting to autosave eslint
