+++ src/components/Delegates/DelegateCardList/DelegateCardList.module.scss
@@ -1,0 +1,50 @@
+@import "@/styles/variables";
+
+.infinite_scroll {
+  display: grid;
+  grid-auto-flow: row;
+  justify-content: space-between;
+  padding-top: $spacing-4;
+  grid-template-columns: repeat(3, 23rem);
+  gap: $spacing-8;
+  @media (max-width: $max-width-6xl) {
+    grid-template-columns: repeat(auto-fit, 23rem);
+    justify-content: space-around;
+  }
+  @media (max-width: $max-width-md) {
+    grid-template-columns: 1fr;
+    gap: $spacing-4;
+  }
+  .link {
+    display: flex;
+    flex-direction: column;
+    .link_container {
+      height: 100%;
+      padding: $spacing-6;
+      border-radius: $spacing-3;
+      background: $white;
+      border-width: $spacing-px;
+      border-color: $gray-300;
+      box-shadow: $box-shadow-newDefault;
+      cursor: pointer;
+      .height_100 {
+        height: 100%;
+      }
+      .summary {
+        display: -webkit-box;
+        word-break: break-word;
+        color: #66676b;
+        overflow: hidden;
+        text-overflow: ellipsis;
+        line-clamp: 2;
+        -webkit-line-clamp: 2;
+        -webkit-box-orient: vertical;
+        font-size: $font-size-base;
+        line-height: $line-height-normal;
+      }
+      .flex_1 {
+        flex: 1;
+      }
+    }
+  }
+}
